Renamed compnent files using pascal case to match convention.
