+++ farmdata2_modules/fd2_tabs/fd2_example/fd2_example.info
@@ -4,5 +4,5 @@ core = 7.x
 package = FarmData2
 
 stylesheets[all][] = '../resources/fd2.css'
-scripts[] = '../resources/dropdownWithAllComponent.js'
+scripts[] = '../resources/DropdownWithAllComponent.js'
