+++ farmdata2_modules/fd2_tabs/resources/DropdownWithAllComponent.spec.comp.js
@@ -1,7 +1,7 @@
 import { mount } from '@cypress/vue'
 
 
-var DropdownComp = require("./dropdownWithAllComponent.js")
+var DropdownComp = require("./DropdownWithAllComponent.js")
 var DropdownWithAllComponent = DropdownComp.DropdownWithAllComponent
 
 describe('Field and Crop Dropdowns', () => {
